+++ Projet 3 Courriers/src/model/content/specialletter/SpecialLetter.java
@@ -5,7 +5,7 @@ import model.content.letter.Letter;
 
 public abstract class SpecialLetter<L extends Letter<?>> extends Letter<L> {
 
-	public SpecialLetter(Inhabitant sender, Inhabitant receiver, L content, int cost) {
+	public SpecialLetter(Inhabitant sender, Inhabitant receiver, L content, double cost) {
 		super(sender, receiver, content, cost);
 	}
